+++ frontend/src/components/index.ts
@@ -6,6 +6,7 @@ export { default as Header } from './Header/Header';
 export { default as Heading } from './Heading/Heading';
 export { default as Input } from './Input/Input';
 export { default as Layout } from './Layout/Layout';
+export { default as Modal } from './Modal/Modal';
 export { default as PagingButton } from './PagingButton/PagingButton';
 export { default as SelectList } from './SelectList/SelectList';
 export { default as SnippetEditor } from './SnippetEditor/SnippetEditor';
@@ -17,4 +18,3 @@ export { default as TemplateEdit } from './TemplateEdit/TemplateEdit';
 export { default as TemplateGrid } from './TemplateGrid/TemplateGrid';
 export { default as Text } from './Text/Text';
 export { default as Toast } from './Toast/Toast';
-export { default as Modal } from './Modal/Modal';
