+++ frontend/src/hooks/authentication/index.ts
@@ -1,3 +1,5 @@
 export { useShowPassword } from './useShowPassword';
 export { useSignupForm } from './useSignupForm';
 export { useCheckLoginState } from './useCheckLoginState';
+export { useAuth } from './useAuth';
+export { useLoginForm } from './useLoginForm';
