+++ frontend/src/hooks/authentication/useAuth.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { AuthContext } from '@/contexts/authContext';
+import { AuthContext } from '@/contexts';
 
 export const useAuth = () => {
   const context = useContext(AuthContext);
