+++ frontend/src/hooks/authentication/useLoginForm.ts
@@ -1,8 +1,8 @@
 import { FormEvent } from 'react';
 
-import { useLoginMutation } from '@/queries/authentication/useLoginMutation';
-import { useInputWithValidate } from '../useInputWithValidate';
-import { validateEmail, validatePassword } from './validates';
+import { useInputWithValidate } from '@/hooks/utils';
+import { useLoginMutation } from '@/queries/authentication';
+import { validateEmail, validatePassword } from '@/service';
 
 export const useLoginForm = () => {
   const { mutateAsync } = useLoginMutation();
