+++ frontend/src/hooks/category/index.ts
@@ -1,0 +1,1 @@
+export { useCategory } from './useCategory';
