+++ frontend/src/pages/index.ts
@@ -1,8 +1,8 @@
+export { default as LandingPage } from './LandingPage/LandingPage';
+export { default as LoginPage } from './LoginPage/LoginPage';
 export { default as MyTemplatePage } from './MyTemplatesPage/MyTemplatePage';
+export { default as NotFoundPage } from './NotFoundPage/NotFoundPage';
+export { default as SignupPage } from './SignupPage/SignupPage';
 export { default as TemplateEditPage } from './TemplateEditPage/TemplateEditPage';
 export { default as TemplatePage } from './TemplatePage/TemplatePage';
 export { default as TemplateUploadPage } from './TemplateUploadPage/TemplateUploadPage';
-export { default as SignupPage } from './SignupPage/SignupPage';
-export { default as LoginPage } from './LoginPage/LoginPage';
-export { default as LandingPage } from './LandingPage/LandingPage';
-export { default as NotFoundPage } from './NotFoundPage/NotFoundPage';
