+++ frontend/src/queries/authentication/index.ts
@@ -1,3 +1,5 @@
 export { useCheckEmailQuery } from './useCheckEmailQuery';
 export { useCheckUsernameQuery } from './useCheckUsernameQuery';
 export { useLoginStateQuery } from './useLoginStateQuery';
+export { useLoginMutation } from './useLoginMutation';
+export { useLogoutMutation } from './useLogoutMutation';
