+++ frontend/src/queries/template/index.ts
@@ -1,1 +1,5 @@
 export { useTemplateListQuery } from './useTemplateListQuery';
+export { useTemplateQuery } from './useTemplateQuery';
+export { useTemplateUploadMutation } from './useTemplateUploadMutation';
+export { useTemplateEditMutation } from './useTemplateEditMutation';
+export { useTemplateDeleteMutation } from './useTemplateDeleteMutation';
