+++ frontend/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { formatRelativeTime } from './formatRelativeTime';
 export { getLanguageByFilename } from './getLanguageByFileName';
+export { removeAllWhitespace } from './removeAllWhitespace';
 export { scroll } from './scroll';
